+++ lib/linkhut/accounts/preferences/user_preference.ex
@@ -13,11 +13,12 @@ defmodule Linkhut.Accounts.Preferences.UserPreference do
     field :show_url, :boolean, default: true
     field :show_exact_dates, :boolean, default: false
     field :default_private, :boolean, default: false
+    field :strip_tracking_params, :boolean, default: false
 
     timestamps(type: :utc_datetime)
   end
 
-  @cast_fields [:timezone, :show_url, :show_exact_dates, :default_private]
+  @cast_fields [:timezone, :show_url, :show_exact_dates, :default_private, :strip_tracking_params]
 
   @doc false
   def changeset(preference, attrs) do