+++ priv/repo/migrations/20260315122653_add_strip_tracking_params_to_user_preferences.exs
@@ -1,0 +1,9 @@
+defmodule Linkhut.Repo.Migrations.AddStripTrackingParamsToUserPreferences do
+  use Ecto.Migration
+
+  def change do
+    alter table(:user_preferences) do
+      add :strip_tracking_params, :boolean, null: false, default: false
+    end
+  end
+end